archive/telegram-bot-legacy/requirements.txt
python-telegram-bot>=21.6,<23
httpx>=0.27,<0.29
python-dotenv>=1.0,<2
python-logging-loki>=0.3,<0.4
